requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
gremlinpython==3.4.4
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.3
pytz==2019.3
requests==2.22.0
sentry-sdk==0.13.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
